api: Add cgroup v2 support for cgclassify
This patch modifies cgroup_attach_task_pid() to check the cgroup
version of the controller to which we are attempting to attach
a process.  If the controller is mounted as a cgroup v1 controller,
cgroup_attach_task_pid() uses the "tasks" file as before.  If
the controller is mounted as v2, it uses "cgroup.procs"

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>