wrapper.c: Store the version in struct cgroup_controller
Store the cgroup version in struct cgroup_controller.  Each
controller can be mounted separately, and thus the cgroup
version is applicable at the controller level.

Note that storing the version in the controller requires
cg_init() to have been called.  cgclassify and cgconfig
needed to be modified slightly to ensure that cg_init()
was invoked prior to creating controllers.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>