cgxget: Introduce cgxget
Add a new libcgroup tool - cgxget.  cgxget is based upon cgget,
but it supports converting from cgroup version to another.

For example, a request like the following will work on a system
running the cgroup v1 cpu controller or the cgroup v2 cpu controller.
	$ cgxget -2 cpu.weight MyCgroup

The return value to the user will match the cgroup version requested.
If the user requests a v1 setting, the return value will be a cgroup
v1 setting.  If the user requests a v2 setting, the return value will
be a cgroup v2 setting.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>