api.c: Simplify subtree control logic in _cgroup_create_cgroup()
Prior to this change _cgroup_create_cgroup() wrote to the subtree
control files in two separate locations - one was recursive starting
at the parent, and one wrote directly to the cgroup being created.

Simplify this logic by recursively writing to the new cgroup and
working the way up through its parents.  This change is intended to
be non-functional and purely code cleanup.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>