ftests: Rework consts directory
Rework the consts directory to make it easier on the users of consts.

Consts class inherits from ConstsCommon.  It also declares a distro
variable in its __init__() that allows users to utilize any consts that
can vary from distro to distro.

The distro classes instantiate the abstract distro class, so that they
are forced to implement the necessary methods.

Ultimately this means that the end user only needs to include the Consts
class.  If they need distro info, they can then use consts.distro.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>